validator: distinction between field and submission
We must distinguish consistently between:
- input field
- submitted value/data by the user

The term "form validation" can be very misleading in a manual.
We do not validate (form) fields but the data in them, submitted by
the user.

That fact that the module(s) are called "input" leads to confusion
as well: where do we mean the module (transitive) and where the
user input (intransitive, received from the pov of the module).

Spelling, grammar, replacement of 8bit chars and other fixes
as we go along.